requirements.txt
streamlit
python-dotenv
google-genai
pandas
